java-dest: add DummyTextDestination to the jar file to make the testing easier
